Create SteamRoller
This is a project I worked on in my free time to make a randomizer. I built a box with a nixie tube display to output random numbers. This is the software component of that project. It it housed on a Raspberry Pi and controls the nixie tube output, a Geiger counter for generating random numbers, two rotary switches, and a button. 

This code listens for a button push, reads the state of the rotary switches, generates a number from the random binary (based on the state of the switches), and outputs the results on a pair of nixie tubes. In addition, it writes 1's and 0's to a binary file every time the Geiger counter detects the radioactive decay of some uranium glass inside the box.