VBA - Execute and Export
This VBA code runs a process that combines data from a SQL database and a Oracle database to get the current balance that all pharmacies owe Humana. It then does some calculations with the data and exports it to an Excel file. Then the code formats the Excel file (adds colors, data formats, etc..) to be used by our outreach team in determining which pharmacies to target with collections efforts.